collage_maker/api/requirements.txt
flask==1.1.1
flask-cors==3.0.8
uwsgi==2.0.18
pyzmq==18.1.0